Added enum EAuctionType and refactored a little bit
