+++ auction-app/src/main/java/com/auctionapp/model/auction/AuctionDTO.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Setter
 public class AuctionDTO {
     private Long id;
-    private String type;
+    private EAuctionType type;
     private Date startTime;
     private Date endTime;
     private Long userId;
