+++ auction-app/src/main/java/com/auctionapp/model/auction/EAuctionType.java
@@ -1,0 +1,6 @@
+package com.auctionapp.model.auction;
+
+public enum EAuctionType {
+    STANDARD,
+    RESERVE
+}
